2주차 과제 1 - Counter 앱 만들고 파일 분리하기
- 컴포넌트 분리
