+++ src/CountButtons.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Button from './Button';
+
+function CountButtons({ onClick }) {
+  return (
+    <p>
+      {[1, 2, 3, 4, 5].map((i) => (
+        <Button type="button" key={i} onClick={() => onClick(i)}>{i}</Button>
+      ))}
+    </p>
+  );
+}
+
+export default CountButtons;
